batctl: Remove redundant check_root_or_die()
Without running as root, batctl would abort. However, this check has
zero value as it is redundant. The kernel will warn us if we don't
have permission to manipulate networking via netlink:

    $ ./batctl if create ra BATMAN_V
    Error - failed to add create batman-adv interface: Operation not permitted

Without this check, it is possible to change networking settings
without root with the CAP_NET_ADMIN permission:

    $ sudo setcap cap_net_admin+eip ./batctl
    $ ./batctl if create ra BATMAN_V
    $ ./batctl if add enp6s0
    $ ./batctl if
    enp6s0: active

Thus, do not force batctl to run as root. Let the system admin and or
user decide how to manage permissions. For those that run batctl as
root, nothing changes.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>
Signed-off-by: Sven Eckelmann <sven@narfation.org>