batctl: Avoid memory leak in print_routing_algos
The opts.remaining_header string is alocated before the netlink callback
object is created. But the callback object allocation can fail and the
function will return in this case. To fix this, either the string buffer
must be freed in this case or the opts.remaining_header allocation can
simply be moved to a later point.

Fixes: 0a14f8800dac ("batctl: Handle nl_cb_alloc errors")
Signed-off-by: Sven Eckelmann <sven@narfation.org>